Move SegmentedControl to separate class
